Training.py
In this section, we load the MNIST dataset and train a simple Convolutional Neural Network (CNN) for digit classification.

After training, we apply ternary quantization to the weights of the model. Ternarization involves mapping the weights to one of three values: [1 0 -1], using a global threshold derived from the maximum absolute weight value. This reduces the model’s precision and size, making it suitable for deployment on low-resource embedded devices.

Finally, we convert this ternarized Keras model into a TensorFlow Lite model, which can be used for efficient inference on edge devices like microcontrollers.